Python/libs/my_app/requirements.txt
streamlit
pandas
numpy
openpyxl
matplotlib
plotly
google.generativeai
sklearn
prophet
